[BZ-1110277] Log event details when addEventData throws an exception
When EventManagerBean.addEventData throws an exception due to a database
or other failure, it is not always clear what the content was around the
failure. With this change, if debug logging is enabled, EventSource
information is logged to make it possible to figure out which resource
and corresponding log file may have been associated with the failure.
Additionally, if trace logging is enabled, the complete Event objects
associated with the EventSource are also logged.